shared-storage/shared-storage-writable-fetch-request-for-data-url.tentative.https.sub.html

<!doctype html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script>
    'use strict';

    promise_test(async t => {
       const innerCode =
         `window.parent.postMessage({fetchStatus: "success"}, '*');`;
       const dataURL = 'data:text/javascript;base64,'
         + btoa(unescape(encodeURIComponent(innerCode)));
        try {
          const response = await fetch(dataURL, {sharedStorageWritable: true});
        } catch (error) {
          assert_equals(error.message,
                        "Failed to execute 'fetch' on 'Window': "
                        + "sharedStorageWritable: sharedStorage operations are "
                        + "not available for opaque origins.");
          return;
        }
        assert_unreached('sharedStorageWritable fetch for data URL did not throw');

    }, 'shared storage fetch request disallowed for data URL');
  </script>
</body>
